refactor structure and pyproject.toml
