+++ pyproject.toml
@@ -20,6 +20,18 @@ classifiers = [
 keywords = [ "Food", "Fusion", "AI"]
 requires-python = ">=3.10"
 dependencies = [
+    "fastapi==0.115.0",
+    "pydantic==2.9.2",
+    "uvicorn==0.30.6",
+    "pyyaml==6.0.2",
+    "python-dotenv==1.0.1",
+    "bcrypt==4.2.0",
+    # -- Hugging face --
+    "transformers==4.44.2",
+    "datasets==3.0.0",
+    # ------------------
+    "numpy==2.1.1",
+    "pandas==2.2.3",
 ]
 
 [tool.coverage.run]
